pyproject.toml
[project]
name = "agent-development-kit-crash-course"
version = "0.1.0"
description = "Examples for learning Google's Agent Development Kit (ADK)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk==1.13.0",
    "yfinance==0.2.65",
    "psutil==7.0.0",
    "litellm==1.76.2",
    "google-generativeai==0.8.5",
    "python-dotenv==1.1.1",
    "deprecated==1.2.18",
]

[project.urls]
Homepage = "https://github.com/google/adk-docs"
Documentation = "https://google.github.io/adk-docs/get-started/quickstart"

[dependency-groups.dev]
dev-dependencies = []
